+++ dotnet/src/webdriver/WebDriver.csproj.prebuild.sh
@@ -1,0 +1,18 @@
+#!/bin/bash
+if [[ ! -f "$1../../../bazel-bin/javascript/webdriver/atoms/get-attribute.js" ]]
+then
+  echo "Building getAttribute atom"
+  bazel build //javascript/webdriver/atoms:get-attribute.js
+fi
+
+if [[ ! -f "$1../../../bazel-bin/javascript/atoms/fragments/is-displayed.js" ]]
+then
+  echo "Building isDisplayed atom"
+  bazel build //javascript/atoms/fragments:is-displayed.js
+fi
+
+if [[ ! -f "$1../../../bazel-bin/javascript/atoms/fragments/find-elements.js" ]]
+then
+  echo "Building findElements atom"
+  bazel build //javascript/atoms/fragments:find-elements.js
+fi