+++ java/client/test/org/openqa/selenium/testing/JUnit4TestBase.java
@@ -67,7 +67,7 @@ public abstract class JUnit4TestBase {
 
   @Before
   public void prepareEnvironment() {
-    environment = GlobalTestEnvironment.get(InProcessTestEnvironment.class);
+    environment = GlobalTestEnvironment.getOrCreate(InProcessTestEnvironment::new);
     appServer = environment.getAppServer();
 
     pages = new Pages(appServer);
@@ -84,7 +84,6 @@ public abstract class JUnit4TestBase {
   @Rule
   public TestRule chain = RuleChain
     .outerRule(new TraceMethodNameRule())
-    .around(new NotificationRule())
     .around(new ManageDriverRule())
     .around(new SwitchToTopRule())
     .around(new NotYetImplementedRule());