+++ java/client/test/org/openqa/selenium/testing/StaticResources.java
@@ -17,6 +17,8 @@
 
 package org.openqa.selenium.testing;
 
+import static org.openqa.selenium.build.DevMode.isInDevMode;
+
 import org.openqa.selenium.build.BazelBuild;
 import org.openqa.selenium.build.InProject;
 
@@ -24,10 +26,20 @@ import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.logging.Logger;
 
 class StaticResources {
+  private static Logger log = Logger.getLogger(StaticResources.class.getName());
 
   static void ensureAvailable() {
+    if (!isInDevMode()) {
+      // we should only need to do this when we're in dev mode
+      // when running in a test suite, our dependencies should already
+      // be listed.
+      log.info("Not in dev mode. Ignoring attempt to copy static resources");
+      return;
+    }
+
     if (!Files.exists(InProject.findProjectRoot().resolve("Rakefile"))) {
       // we're not in dev mode
       return;