+++ java/server/src/org/openqa/selenium/grid/TemplateGridCommand.java
@@ -21,7 +21,9 @@ import com.beust.jcommander.JCommander;
 import com.beust.jcommander.ParameterException;
 import com.beust.jcommander.internal.DefaultConsole;
 
+import com.google.common.collect.Sets;
 import org.openqa.selenium.cli.CliCommand;
+import org.openqa.selenium.grid.config.HasRoles;
 import org.openqa.selenium.grid.config.AnnotatedConfig;
 import org.openqa.selenium.grid.config.CompoundConfig;
 import org.openqa.selenium.grid.config.ConcatenatingConfig;
@@ -33,21 +35,29 @@ import org.openqa.selenium.grid.server.HelpFlags;
 
 import java.io.PrintStream;
 import java.util.LinkedHashSet;
+import java.util.ServiceLoader;
 import java.util.Set;
 
 public abstract class TemplateGridCommand implements CliCommand {
 
   @Override
   public final Executable configure(PrintStream out, PrintStream err, String... args) {
-    Set<Object> allFlags = getFlagObjects();
-
     HelpFlags helpFlags = new HelpFlags();
     ConfigFlags configFlags = new ConfigFlags();
 
-    JCommander.Builder builder = JCommander.newBuilder()
-      .programName(getName())
-      .addObject(configFlags)
-      .addObject(helpFlags);
+    Set<Object> allFlags = new LinkedHashSet<>();
+
+    allFlags.add(helpFlags);
+    allFlags.add(configFlags);
+
+    ServiceLoader.load(HasRoles.class).stream()
+      .map(ServiceLoader.Provider::get)
+      .filter(flags -> !Sets.intersection(getConfigurableRoles(), flags.getRoles()).isEmpty())
+      .forEach(allFlags::add);
+
+    allFlags.addAll(getFlagObjects());
+
+    JCommander.Builder builder = JCommander.newBuilder().programName(getName());
     allFlags.forEach(builder::addObject);
     JCommander commander = builder.build();
     commander.setConsole(new DefaultConsole(out));
@@ -84,8 +94,6 @@ public abstract class TemplateGridCommand implements CliCommand {
       execute(config);
     };
   }
-
-  protected abstract Set<Object> getFlagObjects();
 
   protected abstract String getSystemPropertiesConfigPrefix();