+++ java/server/src/org/openqa/selenium/grid/sessionmap/httpd/SessionMapServer.java
@@ -24,9 +24,8 @@ import org.openqa.selenium.BuildInfo;
 import org.openqa.selenium.cli.CliCommand;
 import org.openqa.selenium.grid.TemplateGridCommand;
 import org.openqa.selenium.grid.config.Config;
-import org.openqa.selenium.grid.server.BaseServerFlags;
+import org.openqa.selenium.grid.config.Role;
 import org.openqa.selenium.grid.server.BaseServerOptions;
-import org.openqa.selenium.grid.server.EventBusFlags;
 import org.openqa.selenium.grid.server.Server;
 import org.openqa.selenium.grid.sessionmap.SessionMap;
 import org.openqa.selenium.grid.sessionmap.config.SessionMapOptions;
@@ -34,9 +33,12 @@ import org.openqa.selenium.netty.server.NettyServer;
 import org.openqa.selenium.remote.http.HttpResponse;
 import org.openqa.selenium.remote.http.Route;
 
+import java.util.Collections;
 import java.util.Set;
 import java.util.logging.Logger;
 
+import static org.openqa.selenium.grid.config.StandardGridRoles.EVENT_BUS_ROLE;
+import static org.openqa.selenium.grid.config.StandardGridRoles.HTTPD_ROLE;
 import static org.openqa.selenium.json.Json.JSON_UTF_8;
 import static org.openqa.selenium.remote.http.Contents.asJson;
 import static org.openqa.selenium.remote.http.Route.get;
@@ -57,10 +59,13 @@ public class SessionMapServer extends TemplateGridCommand {
   }
 
   @Override
-  protected Set<Object> getFlagObjects() {
-    return ImmutableSet.of(
-      new BaseServerFlags(),
-      new EventBusFlags());
+  public Set<Role> getConfigurableRoles() {
+    return ImmutableSet.of(EVENT_BUS_ROLE, HTTPD_ROLE);
+  }
+
+  @Override
+  public Set<Object> getFlagObjects() {
+    return Collections.emptySet();
   }
 
   @Override