+++ java/server/src/org/openqa/selenium/remote/server/ActiveSessions.java
@@ -38,7 +38,7 @@ import java.util.logging.Logger;
  * All current {@link ActiveSession}s that we're aware of.
  */
 public class ActiveSessions {
-  private final static Logger LOG = Logger.getLogger(ActiveSessions.class.getName());
+  private static final Logger LOG = Logger.getLogger(ActiveSessions.class.getName());
 
   private final Cache<SessionId, ActiveSession> allSessions;
   private final List<ActiveSessionListener> listeners = new LinkedList<>();