+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -7,6 +7,7 @@
     - setEdgeChromium
     - setScrollBehavior
     - Supports adding multiple command-line switches using options.addArguments()
+* Add relative locators
 
 ### API Changes